% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cable.R
\name{cable}
\alias{cable}
\title{R6 Class Representing a Cable}
\value{
a data frame with the structure of the cable.
}
\description{
The class cable defines various fields and methods that allows three basic operations (methods):
\itemize{
\item Create a cable structure: given a group type and a group quantity (see the definitions of the fields) create and numerate the conductors of the cable.
\item provide a data frame with the structure of the cable to be used on other process.
\item define other attributes.
}
}
\details{
The object store information about cables to be used in an automation project
}
\examples{

## ------------------------------------------------
## Method `cable$generate_structure`
## ------------------------------------------------

# example code

New_Cable <- cable$new("cable-001")

# Structure definition of a cable grouped by pairs, with 3 pairs,
# individual and overall shield.

New_Cable$generate_structure("pair", 3, TRUE, TRUE)
New_Cable$df_structure()

## ------------------------------------------------
## Method `cable$df_structure`
## ------------------------------------------------

# Conversion of cable structure to a data frame
New_Cable<- cable$new("cable-001")

df_struct <- New_Cable$df_structure()
str(df_struct)

## ------------------------------------------------
## Method `cable$update_attr`
## ------------------------------------------------

# Update of atributtes of a cable

New_Cable <- cable$new("cable-001")

at <- list(
  cable_spec = "cab-spec-001",
  ovr_diam = 0.58,
  ovr_diam_UOM = "in")

New_Cable$update_attr(at)
New_Cable
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tag}}{Cable name: character.}

\item{\code{cable_spec}}{Cable Specification document code: character.}

\item{\code{group}}{Cable agroupation: "cond" per conductor, "pair" per pairs, "triad" per triads, "quad" per quartet: character.}

\item{\code{groups_qty}}{Groups quantity: numeric, int.}

\item{\code{ovr_shield}}{If cable have overall shield or not: logic.}

\item{\code{ind_shields}}{If each group have shield or not: logic.}

\item{\code{ovr_diam}}{The overall diameter of the cable: numeric, float.}

\item{\code{ovr_diam_UOM}}{The overall diameter units of measurement.}

\item{\code{struc}}{Data structure for conductors and groups.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cable-new}{\code{cable$new()}}
\item \href{#method-cable-generate_structure}{\code{cable$generate_structure()}}
\item \href{#method-cable-df_structure}{\code{cable$df_structure()}}
\item \href{#method-cable-update_attr}{\code{cable$update_attr()}}
\item \href{#method-cable-print}{\code{cable$print()}}
\item \href{#method-cable-clone}{\code{cable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cable-new"></a>}}
\if{latex}{\out{\hypertarget{method-cable-new}{}}}
\subsection{Method \code{new()}}{
Create a new cable object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cable$new(tag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{cable tag.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{cable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cable-generate_structure"></a>}}
\if{latex}{\out{\hypertarget{method-cable-generate_structure}{}}}
\subsection{Method \code{generate_structure()}}{
Generates the data structure for conductors and groups, modifies the object updating the 'struc' field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cable$generate_structure(group, groups_qty, ind_shields, ovr_shield)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{Cable agroupation: "cond" per conductor, "pair" per pairs, "triad" per triads, "quad" per quartet: character.}

\item{\code{groups_qty}}{Groups quantity: numeric, int.}

\item{\code{ind_shields}}{If each group have shield or not: logic.}

\item{\code{ovr_shield}}{If cable have overall shield or not: logic.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# example code

New_Cable <- cable$new("cable-001")

# Structure definition of a cable grouped by pairs, with 3 pairs,
# individual and overall shield.

New_Cable$generate_structure("pair", 3, TRUE, TRUE)
New_Cable$df_structure()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cable-df_structure"></a>}}
\if{latex}{\out{\hypertarget{method-cable-df_structure}{}}}
\subsection{Method \code{df_structure()}}{
Converts the 'struc' field to a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cable$df_structure()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Conversion of cable structure to a data frame
New_Cable<- cable$new("cable-001")

df_struct <- New_Cable$df_structure()
str(df_struct)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cable-update_attr"></a>}}
\if{latex}{\out{\hypertarget{method-cable-update_attr}{}}}
\subsection{Method \code{update_attr()}}{
Updates attributes for cable, specifically cable specification, overall diameter, overall diameter UOM, this data is encapsulated on a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cable$update_attr(attrib)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attrib}}{list with the following named fields: cable_spec, ovr_diam, ovr_diam_UOM.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Update of atributtes of a cable

New_Cable <- cable$new("cable-001")

at <- list(
  cable_spec = "cab-spec-001",
  ovr_diam = 0.58,
  ovr_diam_UOM = "in")

New_Cable$update_attr(at)
New_Cable
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cable-print"></a>}}
\if{latex}{\out{\hypertarget{method-cable-print}{}}}
\subsection{Method \code{print()}}{
print the information for cable objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cable$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-cable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
