% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SG.R
\name{SG}
\alias{SG}
\title{R6 Class Representing a Signal}
\description{
The class 'SG' defines various fields and methods for signals that allows:
\itemize{
\item Create a signal: given a name, a device (origin), a control system
(destiny) and a signal type.
\item Assign the signal to a cable.
\item Check that the signal have a complete path from origin to destiny.
}
}
\details{
The SG objects are defined to follow the signal that will be generated on
field devices or control systems.

Each signal should be associated with an origin and a destiny. Additionally,
the signal would be associated to one or various cables that will be conduct
the signal from the origin to the destiny.

As convention, the origin is the device or control system that generates the
signal, the destiny is the device or control system that receives the signal.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tag}}{The name of the signal, character}

\item{\code{origin}}{Origin (device or CS that generates the
signal), character}

\item{\code{destiny}}{Destiny (device or CS that receive the
signal), character}

\item{\code{type}}{Type of the signal (Analog or digital),
character}

\item{\code{path}}{Path that follows the signal from origin to
destiny, list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SG-clone}{\code{SG$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SG-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SG-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SG$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
